+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from "./components.js";
 export * from "./actions.js";
 export * from "./autosave.js";
+export * from "./form-validity.js";
 export * from "./notifications.js";
 export * from "./number.js";
 export * from "./plots.js";